+++ config/routes.rb
@@ -20,6 +20,7 @@ Rails.application.routes.draw do
 
     resources :adverse_events, path: "adverse-events" do
       member do
+        get :dossier
         get :history
         get :log
       end